removed "core" directory and moved contents to common to remove ambiguity of having two similar shared folders
