+++ usaspending_api/common/elasticsearch/client.py
