+++ usaspending_api/recipient/v2/views/list_recipients.py
@@ -7,9 +7,9 @@ from django.db.models import F, Q
 from usaspending_api.common.cache_decorator import cache_response
 from usaspending_api.common.helpers.generic_helper import get_pagination_metadata
 from usaspending_api.common.views import APIDocumentationView
-from usaspending_api.core.validator.pagination import PAGINATION
-from usaspending_api.core.validator.tinyshield import TinyShield
-from usaspending_api.core.validator.utils import update_model_in_list
+from usaspending_api.common.validator.pagination import PAGINATION
+from usaspending_api.common.validator.tinyshield import TinyShield
+from usaspending_api.common.validator.utils import update_model_in_list
 from usaspending_api.recipient.models import RecipientProfile
 from usaspending_api.recipient.v2.lookups import SPECIAL_CASES
